requirements.txt
fake-useragent
bs4
aiohttp
python-dotenv
aiogram
pycryptodome
google-generativeai
sqlalchemy
cryptography
aiosqlite
redis
pydantic-settings
pydantic

